+++ client/src/io/worxbend/gitea4s/http/GiteaEndpoint.scala
@@ -393,6 +393,20 @@ object GiteaEndpoints:
       response = "#/responses/empty"
     )
 
+  val issueEditIssueDeadline: GiteaEndpoint =
+    GiteaEndpoint(
+      method = "POST",
+      path = "/repos/{owner}/{repo}/issues/{index}/deadline",
+      operationId = "issueEditIssueDeadline",
+      parameters = List(
+        GiteaParameter("owner", "path", required = true),
+        GiteaParameter("repo", "path", required = true),
+        GiteaParameter("index", "path", required = true),
+        GiteaParameter("body", "body", required = false)
+      ),
+      response = "#/responses/IssueDeadline"
+    )
+
   val notifyGetList: GiteaEndpoint =
     GiteaEndpoint(
       method = "GET",